Refinement in generate_taxa_indiv_boxplot_long Function
This commit introduces a fine-tuning in the generate_taxa_indiv_boxplot_long function to ensure consistency and correctness in naming and returning plot lists.

Changes Made:
1. The naming of the returned plot_list is aligned with feature.level, enhancing the clarity and understanding of the output. This modification aids in better identification and handling of the output plot list as it now clearly resonates with the feature levels used in the function.
2. A new variable plot_list_all is introduced to collectively hold and return all the plot lists generated in the function. This ensures that all the plots are neatly organized and returned as a single object, improving the manageability and accessibility of the output plots.

By ensuring clear and consistent naming and organization of output plot lists, this update enhances the usability and comprehension of the generate_taxa_indiv_boxplot_long function, leading to a more efficient and effortless user experience.