Update Package Description and URL
This commit updates the `Description` field in the package documentation for MicrobiomeStat. The new description emphasizes the package’s specialized utility for longitudinal microbiome and multi-omics data analysis, including the MicrobiomeStat One Click feature for generating professional microbiome analysis reports with a single click. Additionally, the URL has been updated to include the 'https://' protocol, enhancing the accessibility and security of the website link in the package documentation.