Improved the Functionality and Readability of Generate_beta_ordination_single
This push correspondence to numerous enhancements that have been made in generate_beta_ordination_single for improved ease of use and functionality.

The code has been updated and optimized to make it more readable and efficient. The replication of some arguments such as dist.obj and pc.obj within the function and conditional clause were revisited and refactored reducing redundancy.

Also, changes have been made on how the dataframe is manipulated before plotting: instead of using base R functions for manipulating the dataframe, now using dplyr to filter and rename for improved efficiency and easy debugging.

In addition, this push removed arguments from generate_beta_ordination_single that were previously defined, but never used within the function. This will help to make the code cleaner and easier to understand.

Lastly, adjustments in the manner the plot is built have been done using ggplot2 to make it more user-friendly and interactive. This has been done keeping in mind the anticipated end user interactions with the script.

Overall, while executing the same operation as before, this revised function takes advantage of more efficient coding conventions making it more maintainable and accessible.