Condition Check Fix in Aggregated Taxonomies
This push addressed the issue of incorrect condition checking in the code related to the aggregation of taxonomies. Previously, the check was only confirming if the variable 'agg_taxonomies' was not 'NA'. The updated code now correctly checks if there is any 'NA' value in the 'agg_taxonomies' variable. This fix ensures proper aggregation of taxonomies, even when 'NA' values are present.