Update README with Chinese and Spanish Translations
In this commit, I've added Chinese and Spanish versions of the README for MicrobiomeStat. Our aim is to foster better communication and understanding among our diverse user base. These translations ensure that our resources are more accessible to a broader audience. Feedback on the translations is welcomed to ensure accuracy and clarity.