Updated MicrobiomeStat GitHub Pages with pkgdown::build_site
This commit includes an update to the MicrobiomeStat GitHub Pages, which was carried out using the pkgdown::build_site function. The update aims to improve the quality and accessibility of the documentation, making it easier for users to understand and utilize the MicrobiomeStat package. As always, we welcome feedback and contributions from the community to further enhance the utility and user-friendliness of MicrobiomeStat.