Enhanced Theme Handling and Documentation in Functions
This update brings a significant enhancement to the way themes are handled across our suite of functions. Key changes include the centralization of theme selection through the introduction of the `mStat_get_theme` function and comprehensive updates to the function documentations, ensuring greater clarity and ease of use for users.

**Major Enhancements:**
1. **Centralized Theme Selection:** Implemented the `mStat_get_theme` function across all relevant functions. This change replaces the previously used switch-case theme selection method, streamlining the process and making it more efficient. The `mStat_get_theme` function allows for a cleaner codebase and easier maintenance, especially when adding new themes or modifying existing ones.

2. **Updated Documentation for Theme Parameters:** Enhanced the documentation for the `theme.choice` and `custom.theme` parameters across all affected functions. This update provides users with a detailed guide on how to choose from predefined themes or create their own custom themes. Examples of common customizations for publication-ready figures have been added, making the documentation more practical and user-friendly.

3. **Default Theme Setting:** Standardized the default theme to "bw" (black and white), ensuring a consistent and professional look across all plots, especially useful in formal publications or presentations.

These updates are aimed at improving the user experience by providing clearer options and guidance for theme selection in plotting functions. The addition of practical examples in the documentation further aids users in customizing plots to their specific needs, whether for data exploration or publication purposes.