Avoid automatic rarefaction for 'count' data type
In alignment with our policy not to perform automatic rarefaction on data with `feature.data.type` set to "count", modifications have been made to the diversity analysis functions. The code now checks if the data are already rarefied before proceeding with beta and alpha diversity calculations. If the data are not rarefied, a warning is issued to inform the user. This ensures that rarefaction remains a user-driven process, and they can use the `mStat_rarefy_data` function if rarefaction is deemed necessary.